+++ zebra-chain/src/lib.rs
@@ -35,9 +35,6 @@ pub use parameters::NetworkUpgrade;
 #[cfg(test)]
 use proptest_derive::Arbitrary;
 
-#[cfg(test)]
-pub mod test;
-
 /// An enum describing the possible network choices.
 #[derive(Copy, Clone, Debug, Eq, PartialEq, Hash, Serialize, Deserialize)]
 #[cfg_attr(test, derive(Arbitrary))]
