--- zebra-chain/src/test.rs
@@ -1,2 +1,0 @@
-//! Test support code
-pub mod generate;
